Servicio de correo agregado, patch de activación de usuario modificado con esquema, correción de registro y login. Modificación de "db.js" para que admita múltiple sentencias SQL en una única query.

+++ backend/db.sql
@@ -9,7 +9,7 @@ CREATE TABLE `users` (
   `password` varchar(255) NOT NULL,
   `role` enum("admin","regular") NOT NULL,
   `course` varchar(255) NOT NULL,
-  `dni` mediumint,
+  `dni` int unsigned,
   `has_certificate` bool DEFAULT true,
   `is_active` bool DEFAULT true,
   `last_login` date DEFAULT(now()),
